hack: skip destroying container in workflow runs

+++ packages/apps/session-storage/shopify-app-session-storage-mysql/src/__tests__/mysql.test.ts
@@ -72,7 +72,10 @@ describe('MySQLSessionStorage', () => {
     await storage.disconnect();
     await storage2.disconnect();
 
-    await exec(`podman rm -f ${containerId}`);
+    // eslint-disable-next-line no-process-env
+    if (process.env.IS_WORKFLOW_RUN !== 'true') {
+      await exec(`podman rm -f ${containerId}`);
+    }
   });
 
   const tests = [
